Struct burn_tch::TchTensor

pub struct TchTensor<E: Element, const D: usize> {
    pub tensor: Tensor,
    pub storage: StorageRef,
    /* private fields */
}

A tensor that uses the tch backend.

Fields
tensor: Tensor

Handle to the tensor. Call methods on this field.

storage: StorageRef

The tensor’s storage

Implementations

impl<E: Element, const D: usize> TchTensor<E, D>

pub fn new(tensor: Tensor) -> Self

Create a new tensor.

Note that if the tensor was created from an operation that may reuse the same tensor storage as the parent, you should use from_existing instead.

pub fn from_existing(tensor: Tensor, storage_parent: StorageRef) -> Self

Create a tensor that was created from an operation executed on a parent tensor.

If the child tensor shared the same storage as its parent, it will be cloned, effectively tracking how much tensors point to the same memory space.

pub fn partial(tensor: Tensor, storage_parent: StorageRef) -> Self

Create a tensor that uses a part of its parent tensor such as slice and narrow.

impl<P: Element, const D: usize> TchTensor<P, D>

pub fn mut_ops<F: Fn(&mut Tensor) -> Tensor, EOut: Element, const D_OUT: usize>( &mut self, func: F ) -> Option<TchTensor<EOut, D_OUT>>

Execute an operation on a tensor if the data can be reused.

pub fn unary_ops<FOwn, FRef, EOut: Element, const D_OUT: usize>( self, fown: FOwn, fref: FRef ) -> TchTensor<EOut, D_OUT>where FOwn: Fn(Tensor) -> Tensor, FRef: Fn(&Tensor) -> Tensor,

Execute a unary ops reusing the tensor data if possible.

pub fn binary_ops_tensor<FLMut, FRMut, FRef, EOut: Element, const D_OUT: usize>( lhs: Self, rhs: Self, flmut: FLMut, frmut: FRMut, fref: FRef ) -> TchTensor<EOut, D_OUT>where FLMut: Fn(&mut Tensor, &Tensor) -> Tensor, FRMut: Fn(&Tensor, &mut Tensor) -> Tensor, FRef: Fn(&Tensor, &Tensor) -> Tensor,

Execute a binary ops reusing the tensor data if possible.

impl<E: Element + Default, const D: usize> TchTensor<E, D>

pub fn from_data(data: Data<E, D>, device: Device) -> Self

Creates a new tensor from a shape and a device.

Arguments

data - The tensor’s data.

device - The device on which the tensor will be allocated.

Returns

A new tensor.

impl<E: Element + Default + Copy + Debug, const D: usize> TchTensor<E, D>

pub fn empty(shape: Shape<D>, device: LibTorchDevice) -> Self

Creates an empty tensor from a shape and a device.

Arguments

shape - The shape of the tensor.

device - The device to create the tensor on.

Returns

A new empty tensor.